rewrite slab.c to make it simpler
compileation is ok but not do any test

+++ include/os/slab.h
@@ -1,8 +1,15 @@
+/*
+ * include/os/slab.h
+ *
+ * Created by Le Min (lemin9538@163.com)
+ *
+ */
+
 #ifndef _SLAB_H
 #define _SLAB_H
 
-void *kmalloc(int size, unsigned long flag);
+void *kmalloc(int size, int flag);
 void kfree(void *addr);
-void *kzalloc(int size, unsigned long flag);
+void *kzalloc(int size, int flag);
 
 #endif
